requirements.txt
jupyter=1.0.0
matplotlib=3.1.3
numpy=1.18.1
pandas=1.0.3
python=3.7.6
pytorch=1.4.0
scikit-learn=0.22.1
scipy=1.4.1
seaborn=0.10.0
tensorboardx=2.0
